requirements.txt
requests>=2.28.0
beautifulsoup4>=4.11.0
lxml>=4.9.0
PyMuPDF>=1.21.0
Pillow>=9.0.0
